+++ template/property-grid-footer.php
@@ -143,6 +143,7 @@ function apartmentsync_get_floorplan_info_for_properties_grid( $args ) {
                 $minimum_sqft = get_post_meta( $id, 'minimum_sqft', true );
                 $maximum_sqft = get_post_meta( $id, 'maximum_sqft', true );
                 $available_units = get_post_meta( $id, 'available_units', true );
+                $has_specials = get_post_meta( $id, 'has_specials', true );
                 
                 if ( !isset( $floorplans[$property_id ] ) ) {
                     $floorplans[ $property_id ] = array(
@@ -154,6 +155,7 @@ function apartmentsync_get_floorplan_info_for_properties_grid( $args ) {
                         'minimum_sqft' => array( $minimum_sqft ),
                         'maximum_sqft' => array( $maximum_sqft ),
                         'available_units' => array( $available_units ),
+                        'has_specials' => array( $has_specials ),
                     );
                 } else {
                     $floorplans[ $property_id ]['id'][] = $id;
@@ -164,6 +166,7 @@ function apartmentsync_get_floorplan_info_for_properties_grid( $args ) {
                     $floorplans[ $property_id ]['minimum_sqft'][] = $minimum_sqft;
                     $floorplans[ $property_id ]['maximum_sqft'][] = $maximum_sqft;
                     $floorplans[ $property_id ]['available_units'][] = $available_units;
+                    $floorplans[ $property_id ]['has_specials'][] = $has_specials;
                 }
                 
             endwhile;
@@ -217,7 +220,18 @@ function apartmentsync_get_floorplan_info_for_properties_grid( $args ) {
             $floorplans[$key]['sqftrange'] = $min . '-' . $max;
         }
         
+        // default value
+        $floorplans[$key]['property_has_specials'] = false;
+        
+        // if there are specials, save that
+        $has_specials = $floorplan['has_specials'];
+        
+        if ( in_array( true, $has_specials ) )        
+            $floorplans[$key]['property_has_specials'] = true;
+                        
     }
+    
+    
     
     return $floorplans;
 }